commit final

+++ uvod-u-c-4/prosjek_u_nizu.cpp
@@ -1,13 +1,13 @@
 #include <iostream>
 using namespace std;
 
-int zbrajanje(int broj, int b) {
+float zbrajanje(int broj, int b) {
   return broj+b;
 }
 
 int main(int argc, char *argv[])
 {
-  int a, b, zbroj = 0;
+  float a, b, zbroj = 0;
   cin >> a;
   for (int i = 0;i < a; i++) {
     cin >> b;
